feat(storyPage): use getx and add textScaleFactor
